FUSETOOLS-1704 - Move endpoint specific to a dedicated package
